Fixes #19: contribs mood
The problem here isn't that the contrib column in uram_contribs is NULL;
the problem is that there's a row in the table to have a NULL in.  When
all saliencies are 0.0, there are no contributions to mood, and
uram_contribs should be empty.  (The contribs nbmood command got this
right.)