Fixes #320: buffer overflow fixed
A fixed length character string is used to store the error message for
output when an invalid MGRS string is encountered.

The length of the invalid string was not checked in the Tcl binding
before copying it into the error buffer.  This is now fixed.  A maximum
of 20 characters is output. If the input string is longer than that an
ellipsis is appended to the end of the invalid string.

Retest.